+++ src/main/java/ir/ac/aut/hesabgar/helper/InvoiceHelper.java
@@ -22,8 +22,8 @@ public class InvoiceHelper {
     @Autowired
     private UserInfoRepo userInfoRepo;
 
-    private String findMin(Map<String, Integer> groupBalance) {
-        int min = 0;
+    private String findMin(Map<String, Float> groupBalance) {
+        float min = 0;
         String toBeReturn = null;
         for (String user : groupBalance.keySet()) {
             if (groupBalance.get(user) < min) {
@@ -35,8 +35,8 @@ public class InvoiceHelper {
         return toBeReturn;
     }
 
-    private String findMax(Map<String, Integer> groupBalance) {
-        int max = 0;
+    private String findMax(Map<String, Float> groupBalance) {
+        float max = 0;
         String toBeReturn = null;
         for (String user : groupBalance.keySet()) {
             if (groupBalance.get(user) > max) {
@@ -48,9 +48,9 @@ public class InvoiceHelper {
         return toBeReturn;
     }
 
-    public GroupInfo findPaymentTerm(GroupInfo groupInfo, Map<String, Integer> balance) {
+    public GroupInfo findPaymentTerm(GroupInfo groupInfo, Map<String, Float> balance) {
         List<GroupMember> groupMembers = groupInfo.getMembers();
-        Map<String, Integer> groupBalance = new HashMap<>();
+        Map<String, Float> groupBalance = new HashMap<>();
         for (GroupMember groupMember : groupMembers) {
             groupMember.setPaymentTerms(new ArrayList<>());
         }
@@ -75,8 +75,8 @@ public class InvoiceHelper {
                         }
                     }
 
-                    groupBalance.put(user, 0);
-                    groupBalance.put(user2, 0);
+                    groupBalance.put(user,(float) 0);
+                    groupBalance.put(user2,(float) 0);
                 }
             }
         }
@@ -84,8 +84,8 @@ public class InvoiceHelper {
             String maxDebt = findMin(groupBalance);
             String maxCred = findMax(groupBalance);
             if (groupBalance.get(maxDebt) * -1 == groupBalance.get(maxCred)) {
-                int minAmount = groupBalance.get(maxDebt);
-                int maxAmount = groupBalance.get(maxCred);
+                float minAmount = groupBalance.get(maxDebt);
+                float maxAmount = groupBalance.get(maxCred);
 
                 PaymentTerm paymentTermMaxDebt = new PaymentTerm(maxCred, groupBalance.get(maxDebt));
                 PaymentTerm paymentTermMaxCred = new PaymentTerm(maxDebt, groupBalance.get(maxCred));
@@ -103,8 +103,8 @@ public class InvoiceHelper {
                 groupBalance.put(maxCred, minAmount + maxAmount);
 
             } else if (groupBalance.get(maxDebt) * -1 > groupBalance.get(maxCred)) {
-                int minAmount = groupBalance.get(maxDebt);
-                int maxAmount = groupBalance.get(maxCred);
+                float minAmount = groupBalance.get(maxDebt);
+                float maxAmount = groupBalance.get(maxCred);
 
 
                 PaymentTerm paymentTermMaxDebt = new PaymentTerm(maxCred, maxAmount * -1);
@@ -121,11 +121,11 @@ public class InvoiceHelper {
                 }
 
                 groupBalance.put(maxDebt, minAmount + maxAmount);
-                groupBalance.put(maxCred, 0);
+                groupBalance.put(maxCred, (float) 0);
 
             } else if (groupBalance.get(maxDebt) * -1 < groupBalance.get(maxCred)) {
-                int minAmount = groupBalance.get(maxDebt);
-                int maxAmount = groupBalance.get(maxCred);
+                float minAmount = groupBalance.get(maxDebt);
+                float maxAmount = groupBalance.get(maxCred);
 
 
                 PaymentTerm paymentTermMaxDebt = new PaymentTerm(maxCred, minAmount);
@@ -140,7 +140,7 @@ public class InvoiceHelper {
                         paymentTerms.add(paymentTermMaxDebt);
                     }
                 }
-                groupBalance.put(maxDebt, 0);
+                groupBalance.put(maxDebt,(float) 0);
                 groupBalance.put(maxCred, minAmount + maxAmount);
             }
 
@@ -170,13 +170,13 @@ public class InvoiceHelper {
         }
         addingInvoiceEvent.setEventMembers(eventMembers);
 
-        Map<String, Integer> groupBalance = groupInfo.getGroupBalance();
+        Map<String, Float> groupBalance = groupInfo.getGroupBalance();
         for (String user : addingInvoiceRequest.getGroupShare().keySet()) {
-            int balance = addingInvoiceRequest.getGroupShare().get(user);
+            float balance = addingInvoiceRequest.getGroupShare().get(user);
             groupBalance.put(user, balance + groupBalance.get(user));
         }
 
-        Map<String, Integer> eventBalance = new HashMap<>();
+        Map<String, Float> eventBalance = new HashMap<>();
 
         for (String user : addingInvoiceRequest.getGroupShare().keySet()) {
             eventBalance.put(user, addingInvoiceRequest.getGroupShare().get(user));
@@ -194,10 +194,10 @@ public class InvoiceHelper {
     }
 
     public void updateEventMembersBalance(AddingInvoiceRequest addingInvoiceRequest) {
-        Map<String, Integer> groupShare = addingInvoiceRequest.getGroupShare();
+        Map<String, Float> groupShare = addingInvoiceRequest.getGroupShare();
         for (String userId : groupShare.keySet()) {
             UserInfo userInfo = userInfoRepo.getUserInfoById(userId);
-            int balance = userInfo.getOverallBalance();
+            float balance = userInfo.getOverallBalance();
             userInfo.setOverallBalance(balance + groupShare.get(userId));
             userInfoRepo.save(userInfo);
         }
