+++ src/main/java/ir/ac/aut/hesabgar/request/group/PayingInvoiceRequest.java
@@ -4,7 +4,7 @@ public class PayingInvoiceRequest {
     private String groupId;
     private String debtorUserId;
     private String creditorUserId;
-    private int debtAmount;
+    private float debtAmount;
 
     public String getGroupId() {
         return groupId;
@@ -30,11 +30,11 @@ public class PayingInvoiceRequest {
         this.creditorUserId = creditorUserId;
     }
 
-    public int getDebtAmount() {
+    public float getDebtAmount() {
         return debtAmount;
     }
 
-    public void setDebtAmount(int debtAmount) {
+    public void setDebtAmount(Float debtAmount) {
         this.debtAmount = debtAmount;
     }
 }
